+++ example/lib/textfield/textfield_example.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:widgets_functions/responsive/responsive_widget.dart';
+import 'package:widgets_functions/responsive/responsive_layout.dart';
 import 'package:widgets_functions/text_field/wf_text_field.dart';
 
 class TextFieldExample extends StatefulWidget {
@@ -12,7 +12,7 @@ class TextFieldExample extends StatefulWidget {
 class _TextFieldExampleState extends State<TextFieldExample> {
   @override
   Widget build(BuildContext context) {
-    return ResponsiveWidget(
+    return ResponsiveLayout(
         mobileBody: _mbTextFieldExample(context),
         tabletBody: _mbTextFieldExample(context),
         desktopBody: _mbTextFieldExample(context));